Here's a clear and concise commit message summarizing the changes:
feat: Refactor module initialization and template handling
- Move README and LICENSE copying to a separate module
- Add function to update README content with module name
- Remove unused `copy_and_replace_templates` function
- Update `generate_github_actions_workflow` function to use the new `update_readme_content` function
